requirements.txt
certifi            2025.1.31
charset-normalizer 3.4.1
idna               3.10
markdown-it-py     3.0.0
mdurl              0.1.2
pip                22.0.4
Pygments           2.19.1
python-dotenv      1.0.1
requests           2.32.3
rich               13.9.4
setuptools         58.1.0
tabulate           0.9.0
typing_extensions  4.12.2
urllib3            2.3.0